requirements.txt
ipykernel
python-dotenv
gdown
pandas
streamlit
streamlit_extras
streamlit-ace
plotly
streamlit_authenticator
notion_df
scipy
mplsoccer